Move Skill struct to types.go
